Codec.Candid.TestExports

Description

This modules exports internals soley for the purpose of importing them in the test suite

Documentation

data DidFile

Constructors

DidFile

Fields

defs :: [DidDef TypeName]

service :: DidService TypeName

data TestAssertion

Constructors

CanParse TestInput

CannotParse TestInput

ParseEq Bool TestInput TestInput

data TestInput

Constructors

FromTextual Text

FromBinary ByteString

data CandidTest a

Constructors

CandidTest

Fields

testLine :: Int

testAssertion :: TestAssertion

testType :: [Type a]

testDesc :: Maybe Text

data CandidTestFile

A candid test file

(no support for type definitions yet)

Constructors

CandidTestFile

Fields

testDefs :: [DidDef TypeName]

testTests :: [CandidTest TypeName]

parseCandidTests :: String -> String -> Either String CandidTestFile

Parses a candid spec test file from a string

generateCandidDefs :: [DidDef TypeName] -> Q ([Dec], TypeName -> Q Name)

Turns all candid type definitions into newtypes Used, so far, only in the Candid test suite runner

candidTypeQ :: [Type Name] -> TypeQ

invertHash :: Word32 -> Maybe Text

Inversion of the Candid field label hash